use long for startTxId

+++ hadoop-hdds/common/src/main/java/org/apache/hadoop/hdds/scm/client/ScmClient.java
@@ -380,7 +380,7 @@ public interface ScmClient extends Closeable {
    * @throws IOException
    */
   List<DeletedBlocksTransactionInfo> getFailedDeletedBlockTxn(int count,
-      int startTxId) throws IOException;
+      long startTxId) throws IOException;
 
   /**
    * Reset the failed deleted block retry count.
