+++ hadoop-hdds/framework/src/main/java/org/apache/hadoop/hdds/scm/protocolPB/StorageContainerLocationProtocolClientSideTranslatorPB.java
@@ -718,7 +718,7 @@ public final class StorageContainerLocationProtocolClientSideTranslatorPB
 
   @Override
   public List<DeletedBlocksTransactionInfo> getFailedDeletedBlockTxn(int count,
-      int startTxId) throws IOException {
+      long startTxId) throws IOException {
     GetFailedDeletedBlocksTxnRequestProto request =
         GetFailedDeletedBlocksTxnRequestProto.newBuilder()
             .setCount(count)
