+++ hadoop-hdds/server-scm/src/main/java/org/apache/hadoop/hdds/scm/server/SCMClientProtocolServer.java
@@ -849,7 +849,7 @@ public class SCMClientProtocolServer implements
   }
 
   public List<DeletedBlocksTransactionInfo> getFailedDeletedBlockTxn(int count,
-      int startTxId) throws IOException {
+      long startTxId) throws IOException {
     List<DeletedBlocksTransactionInfo> result;
     try {
       result = scm.getScmBlockManager().getDeletedBlockLog()
